doc: update HIF frame structure
The frame payload description has been deported to its own section, to
clarify that it applies to both native UART and CPC.

Now the 'len' field only describes the frame payload. This way 0 is a
valid value, while before 'len' had to count the mandatory 'fcs' field.

The 'cmd' field has been changed to a byte. If the command space needs
to be expanded in the future, an ID for extended commands will be
allocated.

WI_SUN-2436